+++ remotespark/utils/utils.py
@@ -7,68 +7,16 @@
 
 import os
 import uuid
-import pandas as pd
+
 import numpy as np
-import traceback
+import pandas as pd
 
-from .filesystemreaderwriter import FileSystemReaderWriter
 from .constants import LANG_SCALA, LANG_PYTHON, LANG_R, \
-    SESSION_KIND_SPARKR, SESSION_KIND_SPARK, SESSION_KIND_PYSPARK, INTERNAL_ERROR_MSG, \
-    EXPECTED_ERROR_MSG
-
+    SESSION_KIND_SPARKR, SESSION_KIND_SPARK, SESSION_KIND_PYSPARK
+from .filesystemreaderwriter import FileSystemReaderWriter
 
 first_run = True
 instance_id = None
-
-
-def handle_expected_exceptions(f):
-    """A decorator that handles expected exceptions. Self can be any object with
-    an "ipython_display" attribute.
-    Usage:
-    @handle_expected_exceptions
-    def fn(self, ...):
-        etc..."""
-    import remotespark.livyclientlib.exceptions as e
-    exceptions_to_handle = (e.BadUserDataException, e.LivyUnexpectedStatusException, e.FailedToCreateSqlContextException,
-                            e.HttpClientException, e.LivyClientTimeoutException, e.SessionManagementException)
-
-    # Notice that we're NOT handling e.DataFrameParseException here. That's because DataFrameParseException
-    # is an internal error that suggests something is wrong with LivyClientLib's implementation.
-    def wrapped(self, *args, **kwargs):
-        try:
-            out = f(self, *args, **kwargs)
-        except exceptions_to_handle as err:
-            self.ipython_display.send_error(EXPECTED_ERROR_MSG.format(err))
-            return None
-        else:
-            return out
-    wrapped.__name__ = f.__name__
-    wrapped.__doc__ = f.__doc__
-    return wrapped
-
-
-def wrap_unexpected_exceptions(f, execute_if_error=None):
-    """A decorator that catches all exceptions from the function f and alerts the user about them.
-    Self can be any object with a "logger" attribute and a "ipython_display" attribute.
-    All exceptions are logged as "unexpected" exceptions, and a request is made to the user to file an issue
-    at the Github repository. If there is an error, returns None if execute_if_error is None, or else
-    returns the output of the function execute_if_error.
-    Usage:
-    @wrap_unexpected_exceptions
-    def fn(self, ...):
-        ..etc """
-    def wrapped(self, *args, **kwargs):
-        try:
-            out = f(self, *args, **kwargs)
-        except Exception as e:
-            self.logger.error("ENCOUNTERED AN INTERNAL ERROR: {}\n\tTraceback:\n{}".format(e, traceback.format_exc()))
-            self.ipython_display.send_error(INTERNAL_ERROR_MSG.format(e))
-            return None if execute_if_error is None else execute_if_error()
-        else:
-            return out
-    wrapped.__name__ = f.__name__
-    wrapped.__doc__ = f.__doc__
-    return wrapped
 
 
 def expand_path(path):
